Create Dollars and Cents
Kamal was traveling from Korea to USA and he was not aware of the currency system of USA. Can you please help him to add two dollars and cents?

Note: 1 dollar=100 cents



Input Format

Input consists of 4 integers. The first two inputs correspond to the value of the first dollar and cent. The next two inputs correspond to the value of the second dollar and cent.



Output Format

The output should print the sum of dollars and cents.



Sample Input

30

10

140

99



Sample Input

171

9