Create Number of days
The given are two positive integers Year and Month, the task is to find the number of days corresponding to each month of the given year where 1 is January, 2 is February, 3 is March, and so on.

Note: The Input range of year: Minimum Year = 1900 and Maximum Year = 9999 i.e (1900<=Year<=9999). The Input range of month: Minimum Month = 12 and Maximum Year = 12 i.e (1<=month<=12) If the given year does not lie between the given range, output 0.



Input Format

The input consists of two lines. The first line contains an integer, which is the Year. The second line contains an integer, which is the Month.



Output Format

The output prints the number of days. Refer to the sample output



Sample Input 1

2000

2



Sample Output 1

29 Days