Create Spiral Pattern
Given an integer matrix of size n*n. n is the number of rows and columns. Traverse it in a spiral form.



Input Format

The first line contains N, which represents the number of rows and columns of a matrix.
The next N lines contain N values, each representing the values of the matrix.


Output Format

A single line containing integers with space represents the desired traversal.



Sample Input 1

3

1 2 3

4 5 6

7 8 9



Sample Output 1

1 2 3 6 9 8 7 4 5