Create Sum of rows and columns
Ram has a fruit shop. He has arranged some set of fruits in the column and row-wise. Ram needs to find which row and column have a maximum number of fruits. Help him to find out.



Input Format

Input consists of 2 integers(size of rows and columns) and 1 2D array.



Output Format

The output prints the sum of all rows and columns as well as print the row and column which has the maximum sum. Refer to the sample output.



Sample Input 1

3

3

1 6 8

2 5 1

3 8 2



Sample Output 1

The Sum of rows is 15 8 13

Row 1 has a maximum sum

The Sum of columns is 6 19 11

Column 2 has the maximum sum