Create Remove all Characters in Second String that are present in First String
An online test was conducted for the students for taking up a new course. In the test, the students will be provided with two words. They need to remove the characters in the second word which is present in the first word and have to type the rest. Write a program to remove all characters in the second string which are present in the first string



Input Format

Input consists of two strings.



Output Format

The output consists of one string.



Sample Input 0

Motor

Motorcycle



Sample Output 0

cycle



Explanation

﻿The characters which are not present in the first string but present in the second string are a cycle and hence the output is cycle.